requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11
idna==2.6
Jinja2==2.10
jsonpickle==0.9.5
MarkupSafe==1.0
py-dateutil==2.2
pytz==2017.3
requests==2.18.4
six==1.11.0
